Implemented external JS re-execution
This change now reruns any saved JS files when opening the modal window.
This is because some field types may need to make DOM changes in their
external files.
This also means that libraries are being initialized multiple times, but
hopefully they're smart enough to deal with that. So far, no issues.